# Show deleted_rows_scanned stats in verbose mode