The `deleted_rows_scanned` in QueryStats is usuful for diagnosing bottlenecks in scanned tables with many insertion and deletion.   

I added `deleted rows scanned` section in verbose mode.

### Before 
```
spanner> select count(*) from new_orders;
+--------+
|        |
+--------+
| 835790 |
+--------+
1 rows in set (70.49 msecs)
timestamp: 2021-03-09T14:04:10.376489+09:00
cpu:       325.3 msecs
scanned:   835790 rows
optimizer: 2
```

### After
```
spanner> select count(*) from new_orders;
+--------+
|        |
+--------+
| 835790 |
+--------+
1 rows in set (70.49 msecs)
timestamp:            2021-03-09T14:04:10.376489+09:00
cpu time:             325.3 msecs
rows scanned:         835790 rows
deleted rows scanned: 232750 rows
optimizer version:    2
```